Applied HARMONY-3109 [drlvm][jit][opt] javax.swing.plaf.UIResourceTest fails with Jitrino.OPT, but passes on pure JET
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@502588 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/codegenerator/ia32/Ia32CgUtils.cpp
@@ -127,7 +127,7 @@ bool OpndUtils::isConstAreaItem(const Opnd* op)
 bool OpndUtils::isFPConst(const Opnd* op, double dVal)
 {
     const void* addr = extractAddrOfConst(op);
-    return (addr == NULL) ? false : (dVal == *(const float*)addr);
+    return (addr == NULL) ? false : (dVal == *(const double*)addr);
 }
 
 bool OpndUtils::isFPConst(const Opnd* op, float fVal)
